Add JavaFX dependencies and initial application
Create .editorconfig to specify common formatting information.

Update build.gradle to include JavaFX, osdetector is required to select
the correct JavaFX dependency. Add JUnit5 and JMockit dependencies.

Create settings.gradle to add a workaround for the osdetector plugin
version 1.6.0 not working as Gradle plugin.

Update GridGenerator to extend JavaFX's Application class and implement
start() with a simple window as a default implementation.

Create GridGeneratorLauncher to be used as the main class of the
application, it's main method should launch GridGenerator. The launcher
allows the runtime to no longer require a JavaFX install.

Create GridGeneratorLauncherTest to verify that the program arguments
are passed through to the launched application.

Update GridGeneratorTest to add a test verifying that the window title
is set to "Grid Generator".