Add TravisCI and CodeCov integration
Update build.gradle to add the JaCoCo plugin with some configuration for
coverage report destination folders. HTML and XML reports should both be
enabled.

Update gradle-wrapper.properties to increase the project's Gradle
version to 5.1.1 due to problems with the JaCoCo plugin and Java 11 in
the current Gradle version.

Create .travis.yml with basic configuration for Java with OpenJDK11 and
uploading of code coverage reports to CodeCov.

Update README.md to add badges for the build status and code coverage.