Create simple FXML UI layout with menu bar
Create GridGenerator.fxml containing a BorderPane with a MenuBar in top.
The MenuBar should have File > Exit, Input Participants and Draw Grids
as menu items.

Update GridGenerator to load GridGenerator.fxml as the initial scene.

Update GridGeneratorTest to verify that the menu items are shown as
expected.

Update build.gradle to add dependencies for FXML and TestFX, also set
the test compilation to use UTF-8 encoding to ensure that testing for
pseudo translations is consistent.

Create Labels.properties with default strings for the UI menu items.

Create Labels_en_PSEUDO.properties with pseudo translation for the menu
items, this should be used for testing purposes to verify that the menu
items are internationalized properly.