Create scene for input of a class's participants
Create input-class-participants.fxml representing a scene with
participant input components and a list to display added participants.
Buttons should be included for "Add", "Delete" and "Clear".

Update Labels.properties and associated bundles to add text for the
buttons.

Create InputClassParticipantsController with methods for adding, deleting
and clearing the list of participants which can be called from the
buttons. The buttons should only be enabled when valid input or
selections are made.

Create Messages.properties and associated bundles with the error
messages to be returned by the controller.

Create InputClassParticipantsControllerTest with unit tests for the
various scenarios covered by the controller.

Update build.gradle to add the JUnit params dependency to allow
parameterized tests to be used.