Create controller to handle toolbar menu actions
Create GridGeneratorController as a top level controller with methods to
be called by the toolbar menu items.

Update GridGenerator.fxml to include the new controller and also move
"Input Participants" and "Draw Grids" under a new "Navigate" menu.

Update Labels.properties and associated bundles to include a string for
the navigate menu.

Update GridGeneratorTest to verify and handle the addition of the
navigate menu.