Add a confirmation dialog on clear of participants
When clearing participants a confirmation message should be displayed so
that the user can not clear the list by mistake.

Update InputClassParticipantsController to add a confirmation alert
before clearing participants.

Update Messages.properties and associated bundles to add a message for
the confirmation dialog.

Update InputClassParticipantsControllerTest to verify the change in
behaviour for clearing participants.