Implement the Input Participant menu item's action
When the Input Participant menu item is selected a tab pane should be
displayed with tabs for each class. The tabs will each contain an
instance of input-class-participants.fxml.

Update GridGenerator.fxml to give the BorderPane an fx:id.

Update GridGeneratorController to set the center of the BorderPane to a
TabPane loaded from input-participants.fxml when
displayInputParticipants() is triggered.

Create GridGeneratorControllerTest to verify the behaviour of
GridGeneratorController.

Create input-participants.fxml containing an empty TabPane.

Create InputParticipantsController to create a tab for each class during
initialization. The tabs should be loaded from
input-class-participants.fxml.

Create InputParticipantsControllerTest to verify that the class tabs are
loaded in to the TabPane.