Create a preference helper using Java Preferences
A preference helper is needed to store and retrieve preference values to
be used in the application, such as class lists, participant lists and
validity filters. This is to allow both customization and persisting
input between application executions.

Create PreferenceHelper with getters and setters for the needed
preferences, this should use the Java Preferences API.

Create PreferenceHelperTest to verify that the setters and getters work
and that the default values are correct.

Update InputParticipantsController to get the list of participant
classes to create tabs for from the PreferenceHelper.

Update InputClassParticipantsController to get the participant validator
from the PreferenceHelper.