Use preference to backup and restore participants
When participants are added they should be backed up to a preference,
when the application is ran again the participants should be restored
from the preference.

Update InputClassParticipantsController to perform the backup and
restore functions. The class name to store participants against should
be provided as a parameter to initialize the behaviour.

Update InputParticipantController to initialize the class input data
when initializing the class tabs.

Update Messages.properties and associated bundles to add error messages
for both the backup and restore of participants.