Initialize preference defaults in single place
Currently defaults for preferences are hardcoded to be used when
preference value are not found, instead the preferences should all be
initialized to default values in a single place.

Update PreferenceHelper to add an initializePreferences() method which
will pre-populate all preferences with default values.

Update GridGenerator to call PreferenceHelper's new method so
preferenecs are initialized when the application starts.

Update InputParticipantsController and InputClassParticipantsController
to remove the assumed preference defaults as they should no longer be
required.