Begin implementing draw grid functionality
Create a basic layout for draw grids and display it when selecting the
relevant menu item. The layout should contain a combo box which can be
used to select excluded grids.

Create draw-grids.fxml with a basic layout containing a combobox for
excluded grids.

Update GridGeneratorController to load draw-grids.fxml when Navigate >
Draw Grids is selected from the menu bar.

Create DrawGridsController to implement behaviour to allow a user to
select the excluded grids. The combo box should be populated based on
the number of grids preference.