Implement first part of grid draw
The initial behaviour for drawing grids is randomizing a collection of
participants and spliting them in to as many races as needed. Grids
should be able to be excluded and participants not be assigned to those
grids in any of the generated races.

Update build.gradle to bump the JMockit version to 1.46.

Create GridDrawHelper to perform the grid draw, the randomization should
be based on a seed generated from the list of participants, the excluded
grids and the date.
Create GridDrawHelperTest to verify that the grid draw is as expected
for the different scenarios.