Create preferences for grouping participants
Participants should be groupable based on a RegEx filter, this filter
will be applied if spliting the group out from the participants would
leave more remaining participants than a threshold.

Update PreferenceHelper to add two new preferences, one for setting the
participant grouping filter and the other for setting a threshold for
when the grouping is applied.