Create preference for number of heats to generate
Grid Generator should generate multiple heats at a time, a preference is
needed to allow the number of heats to draw to be specified.

Update PreferenceHelper to add a preference for the number of heats to
draw.

Update GridGeneratorTest and PreferenceHelperTest to verify the
preference is handled properly and has a default initialized value.