Allow generating multiple heats at one time
Update GridDrawHelper to draw multiple heats based on the number of
heats preference value. The return value will now be a list of heats,
containing a list of races for that heat and a list of participants for
each race in grid order. The heat number should be added to the hash
that is used as the randomness seed so allow each heat to give different
results.

Update GridDrawHelper to add two new tests to cover zero and multiple
heat scenarios and to update the expected grids as the random seed was
changed.