Implement display of drawn grids
When opening the Draw Grids layout or modifying the excluded grids the
grids should be regenerated and displayed for each class.

Update draw-grids.fxml to add a TabPane to display class drawn grids
within.

Update DrawGridsController to add class tabs to the TabPane and create a
table per heat with the drawn grids displayed. Add an event listener on
the excluded grids combo box to redraw the grids when the combo is
hidden.

Update Labels.properties and associated bundles to add a new label for
the heat tables.

Update GridDrawHelper to return an empty list of heats if there are no
participants.