Add basic printing functionality to grid draw
Update draw-grids.fxml to add a "Print Current" button.

Update DrawGridsController to add a method for printing the currently
selected class tab. Add additional information on the tab displaying the
meeting name and class to allow easier printing. Change some of the
properties of the heat tables to allow more consistent sizing and
behaviour for printing purposes.