Improve formatting of the heat grid table
Update DrawGridsController to add a new column to the heat table
indicating the race number for each row. The column and table sizing
should also be updated to allow better printing of the tables. Use the
meeting name preference to get the meeting name to display when
printing.
Several sections of the controller's node generation could be taken care
of within FXML, add TODOs for those sections so that work can be done at
a later date.