Perform initial styling of heat tables
The heat tables currently have the default application styling, as these
tables are to be printed the should be styled in a more printer friendly
way.

Create draw-grids.css with styling for the grid printouts.

Update draw-grids.fxml to import the draw-grids.css file.

Update DrawGridsController to rearrange the headers to make it clearer
that columns are grids and rows are races.